misc updates and optimizations
Enemies no longer use the tagging system to find the player's shots or
the player, which speeds things up greatly.  The player is not available
without having to use the tagging system at all, and the player's shots
are now handled by the player instead of using the global objects list,
making them easier to access and everything faster overall.